Created Task entity
Created entity for ORM usage
