+++ src/tasks/dto/get-tasks-filter.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsOptional, IsString } from "class-validator";
-import { TaskStatus } from "../task.model";
+import { TaskStatus } from "../task-status.enum";
 
 export class GetTaskFilterDto {
     // this is my enum data type
