+++ src/tasks/entity/task.entity.ts
@@ -1,0 +1,17 @@
+import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
+import { TaskStatus } from "../task-status.enum";
+
+@Entity()
+export class Task {
+    @PrimaryGeneratedColumn('uuid')
+    id: string;
+
+    @Column()
+    title:string;
+    
+    @Column()
+    description: string;
+
+    @Column()
+    status: TaskStatus;
+}
