+++ src/tasks/tasks.controller.ts
@@ -11,47 +11,47 @@
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
 import { UpdateTaskStatusDto } from './dto/update-task-status.dto';
-import { Task, TaskStatus } from './task.model';
+import { TaskStatus } from './task-status.enum';
 import { TasksService } from './tasks.service';
 
 @Controller('tasks')
 export class TasksController {
   constructor(private tasksService: TasksService) {}
 
-  @Get()
-  getTasks(@Query() fillterDto: GetTaskFilterDto): Task[] {
-    // if any filter is actived, call getTaskWillFilter
-    //else get all tasks
-    if(Object.keys(fillterDto).length){ // if any of the optinal keys was actived. run this function
-      return this.tasksService.getTasksWithFilters(fillterDto);
-    } else {
-      return this.tasksService.getAllTasks();
-    }
+  // @Get()
+  // getTasks(@Query() fillterDto: GetTaskFilterDto): Task[] {
+  //   // if any filter is actived, call getTaskWillFilter
+  //   //else get all tasks
+  //   if(Object.keys(fillterDto).length){ // if any of the optinal keys was actived. run this function
+  //     return this.tasksService.getTasksWithFilters(fillterDto);
+  //   } else {
+  //     return this.tasksService.getAllTasks();
+  //   }
 
-  }
+  // }
 
-  @Get('/:id')
-  getTaskById(@Param('id') id: string): Task {
+  // @Get('/:id')
+  // getTaskById(@Param('id') id: string): Task {
 
-    return this.tasksService.getTaskById(id);
-  }
+  //   return this.tasksService.getTaskById(id);
+  // }
 
-  @Post()
-  createTask(@Body() CreateTaskDto: CreateTaskDto): Task {
-    return this.tasksService.createTask(CreateTaskDto);
-  }
+  // @Post()
+  // createTask(@Body() CreateTaskDto: CreateTaskDto): Task {
+  //   return this.tasksService.createTask(CreateTaskDto);
+  // }
 
-  @Delete('/:id')
-  deleteTask(@Param('id') id: string): void {
-    return this.tasksService.deleteTask(id);
-  }
+  // @Delete('/:id')
+  // deleteTask(@Param('id') id: string): void {
+  //   return this.tasksService.deleteTask(id);
+  // }
 
-  @Patch('/:id/status')
-  updateTaskStatus(
-    @Param('id') id: string,
-    @Body() UpdateTaskStatusDto: UpdateTaskStatusDto,
-  ): Task {
-    const { status } = UpdateTaskStatusDto;
-    return this.tasksService.updateTaskStatus(id, status);
-  }
+  // @Patch('/:id/status')
+  // updateTaskStatus(
+  //   @Param('id') id: string,
+  //   @Body() UpdateTaskStatusDto: UpdateTaskStatusDto,
+  // ): Task {
+  //   const { status } = UpdateTaskStatusDto;
+  //   return this.tasksService.updateTaskStatus(id, status);
+  // }
 }
