+++ src/tasks/tasks.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { TasksController } from './tasks.controller';
 import { TasksService } from './tasks.service';
-
+import { TasksRepository } from './repository/tasks.repository';
 @Module({
+  imports: [TypeOrmModule.forFeature([TasksRepository])],
   controllers: [TasksController],
   providers: [TasksService]
 })
